chore: refactors messenger to follow other zipkin instrumentations.
